✨ feat: 提取部分通用函数到 pp_utils / Extract some of the generic functions to pp_utils
