+++ modules/pp_utils/src/lib.rs
@@ -171,6 +171,79 @@ pub fn find_cover(video: &Path) -> Option<PathBuf> {
     None
 }
 
+/// 获取临时文件目录（优先使用 `PP_EXE_DIR` 环境变量指定目录下的 `tmp` 子目录）。
+/// Get the temporary file directory (prefers a `tmp` subdirectory under `PP_EXE_DIR` env var).
+pub fn tmp_dir() -> PathBuf {
+    let base = env::var("PP_EXE_DIR")
+        .map(PathBuf::from)
+        .unwrap_or_else(|_| {
+            env::current_exe()
+                .ok()
+                .and_then(|p| p.parent().map(|d| d.to_path_buf()))
+                .unwrap_or_else(|| PathBuf::from("."))
+        });
+    let tmp = base.join("tmp");
+    std::fs::create_dir_all(&tmp).ok();
+    tmp
+}
+
+/// 使用 ffprobe 获取图片的宽度和高度。
+/// Get image width and height using ffprobe.
+///
+/// # 返回值 / Returns
+/// `(width, height)`，失败时返回 `None`。
+/// `(width, height)`, or `None` on failure.
+pub fn image_dimensions(path: &Path) -> Option<(u32, u32)> {
+    use std::process::Command;
+    let out = Command::new("ffprobe")
+        .args([
+            "-v", "error",
+            "-select_streams", "v:0",
+            "-show_entries", "stream=width,height",
+            "-of", "csv=p=0",
+        ])
+        .arg(path)
+        .stdout(Stdio::piped())
+        .stderr(Stdio::null())
+        .output()
+        .ok()?;
+    let s = String::from_utf8_lossy(&out.stdout);
+    let mut parts = s.trim().splitn(2, ',');
+    let w: u32 = parts.next()?.trim().parse().ok()?;
+    let h: u32 = parts.next()?.trim().parse().ok()?;
+    Some((w, h))
+}
+
+/// 使用 ffprobe 获取视频的时长、宽度和高度。
+/// Get video duration, width, and height using ffprobe.
+///
+/// # 返回值 / Returns
+/// `(duration_secs, width, height)`，失败时返回 `None`。
+/// `(duration_secs, width, height)`, or `None` on failure.
+pub fn video_meta(input: &Path) -> Option<(f64, i32, i32)> {
+    let out = Command::new("ffprobe")
+        .args([
+            "-v", "error",
+            "-select_streams", "v:0",
+            "-show_entries", "format=duration:stream=width,height",
+            "-of", "csv=p=0",
+        ])
+        .arg(input)
+        .stdout(Stdio::piped())
+        .stderr(Stdio::null())
+        .output()
+        .ok()?;
+    let s = String::from_utf8_lossy(&out.stdout);
+    let mut lines = s.lines().filter(|l| !l.trim().is_empty());
+    let dims_line = lines.next()?;
+    let dur_line = lines.next()?;
+    let mut dims = dims_line.splitn(2, ',');
+    let w: i32 = dims.next()?.trim().parse().ok()?;
+    let h: i32 = dims.next()?.trim().parse().ok()?;
+    let dur: f64 = dur_line.trim().parse().ok()?;
+    Some((dur, w, h))
+}
+
 /// 进度上报的缩放基数（10000 = 100.00%）。
 /// Progress reporting scale base (10000 = 100.00%).
 pub const PROGRESS_SCALE: u32 = 10_000;
